requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Brotli==1.1.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cftime==1.6.3
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.2
conda-pack==0.7.1
contourpy==1.1.1
coverage==7.5.0
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.5
filelock==3.13.1
flake8==7.0.0
fonttools==4.43.1
fsspec==2024.3.1
future==1.0.0
fvcore==0.1.5.post20221221
h11==0.14.0
h5py @ file:///croot/h5py_1691589708553/work
hear21passt==0.0.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.21.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.34.1
importlib-resources==6.1.0
inflate64==1.0.0
iniconfig==2.0.0
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.2
jams==0.3.4
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lmdb @ file:///home/conda/feedstock_root/build_artifacts/python-lmdb_1695497906644/work
lxml==5.1.0
MarkupSafe==2.1.5
mat73 @ file:///home/conda/feedstock_root/build_artifacts/mat73_1693524041622/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mir-eval==0.7
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.7
multivolumefile==0.2.3
mypy==1.10.0
mypy-extensions==1.0.0
netCDF4==1.6.5
numba==0.58.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-ml-py3==7.352.0
opencv-python==4.8.1.78
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487166627/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
pluggy==1.5.0
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.0
pybcj==1.0.2
pybind11==2.12.0
pycodestyle==2.11.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.20.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.1
pyppmd==1.1.0
pyproject==1.3.1
pyroomacoustics==0.7.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pysofaconventions==0.1.5
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.17
python-telegram-bot==21.10
pytz==2024.1
PyYAML==6.0.1
pyzstd==0.15.10
referencing==0.35.0
regex==2023.12.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rpds-py==0.18.0
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1692501296392/work/base/dist/scipy-1.11.2-cp39-cp39-linux_x86_64.whl#sha256=3b1a8484ce148b8f528582054e76a2a71f677f9b8ed2c1871547ec805e38aa03
seaborn==0.13.2
sentencepiece==0.1.99
simpleaudio==1.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
sortedcontainers==2.4.0
soundata==0.1.3
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1676571469739/work
soxr==0.3.7
spatialscaper @ file:///disk6/yxdong/SpatialScaper-main
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
termcolor==2.4.0
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.16
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchinfo==1.8.0
torchvision==0.14.0
tornado==5.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
traitlets==5.14.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
webrtcvad==2.0.10
websockets==14.1
widgetsnbextension==4.0.10
ws4py==0.3.2
yacs==0.1.8
zipp==3.17.0
